% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalise.R
\name{assign_test_control}
\alias{assign_test_control}
\title{Roll up a GA lat-long dataset into test and control areas}
\usage{
assign_test_control(
  ga_regionalised,
  test_region = NULL,
  control_region = NULL,
  pre_start = NULL,
  pre_end = NULL,
  test_start = NULL
)
}
\arguments{
\item{test_start}{}
}
\description{
Roll up a GA lat-long dataset into test and control areas
}
