% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_backup.R
\name{backup_to_cloud}
\alias{backup_to_cloud}
\title{Back up the current GeoxR working directory to Google Cloud}
\usage{
backup_to_cloud(
  project_name = NULL,
  cloud_bucket = "itv_geox_data",
  file_types = c(".rds", ".qmd")
)
}
\description{
Back up the current GeoxR working directory to Google Cloud
}
