% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalise.R
\name{calc_ga_uplifts}
\alias{calc_ga_uplifts}
\title{Calculate an uplifts table from GA lat-long data}
\usage{
calc_ga_uplifts(
  ga,
  test_region = test_region,
  control_region = control_region,
  pre_start = pre_start,
  pre_end = pre_end,
  test_start = test_start
)
}
\arguments{
\item{test_start}{}
}
\description{
Calculate an uplifts table from GA lat-long data
}
