% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalise.R
\name{calculate_geoLifts}
\alias{calculate_geoLifts}
\title{Title}
\usage{
calculate_geoLifts(
  kpi_list = NULL,
  test_regions = NULL,
  exclude_regions = NULL,
  test_start = NULL,
  test_end = NULL,
  benchmark_advertiser = NA,
  benchmark_comment = NA,
  benchmark_schedule = NA
)
}
\arguments{
\item{benchmark_schedule}{}
}
\description{
Title
}
