% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplifts.R
\name{calculate_uplifts}
\alias{calculate_uplifts}
\title{Title}
\usage{
calculate_uplifts(
  ga = NULL,
  test_region = NULL,
  control_region = NULL,
  pre_start = NULL,
  pre_end = NULL,
  test_start = NULL,
  adjust_for_bleed_into_control = TRUE,
  spike_adjustment = 0
)
}
\arguments{
\item{spike_adjustment}{}
}
\description{
Title
}
