% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics4.R
\name{ga4_demographics}
\alias{ga4_demographics}
\title{Title}
\usage{
ga4_demographics(
  ga_id = NULL,
  date_min = NULL,
  date_max = NULL,
  filter_clause = NULL,
  metric = "sessions",
  demographic_selected = "userAgeBracket"
)
}
\arguments{
\item{demographic_selected}{}
}
\description{
Title
}
