% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics4.R
\name{ga4_query}
\alias{ga4_query}
\title{Title}
\usage{
ga4_query(
  ga_id = NULL,
  date_min = NULL,
  date_max = NULL,
  filter_clause = NULL,
  metric = "sessions",
  scale_to_national = TRUE
)
}
\arguments{
\item{scale_to_national}{}
}
\description{
Title
}
