% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics4.R
\name{ga4_summarise_sources}
\alias{ga4_summarise_sources}
\title{Extract total traffic by channel group}
\usage{
ga4_summarise_sources(
  ga_id = NULL,
  date_min = NULL,
  date_max = NULL,
  metric = "sessions"
)
}
\arguments{
\item{metric}{}
}
\description{
Extract total traffic by channel group
}
