% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolift_charts.R
\name{geoLift_plotly_test_control}
\alias{geoLift_plotly_test_control}
\title{Draw a plotly test uplift chart}
\usage{
geoLift_plotly_test_control(
  geoLift,
  data_start,
  test_start,
  title = "",
  ytitle = "Visits"
)
}
\arguments{
\item{geoLift}{Output from GeoLift::GeoLift()}
}
\value{
Plotly chart
}
\description{
Draw a plotly test uplift chart
}
