% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spots_ga4.R
\name{get_adjustment_factor_ga4}
\alias{get_adjustment_factor_ga4}
\title{Title}
\usage{
get_adjustment_factor_ga4(ga_hour, spots, test_region, control_region)
}
\arguments{
\item{control_region}{}
}
\description{
Title
}
