% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics4.R
\name{get_ga4_all}
\alias{get_ga4_all}
\title{Title}
\usage{
get_ga4_all(
  ga_id = NULL,
  date_min = NULL,
  date_max = NULL,
  brand_term = NULL,
  conversion_goal = NULL,
  hour_level = TRUE
)
}
\arguments{
\item{minute_level}{}
}
\description{
Title
}
