% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_template.R
\name{new_geoxR_project}
\alias{new_geoxR_project}
\title{Creates a new GeoX project folder}
\usage{
new_geoxR_project(file_name = NULL, ext_name = "quartotemplate")
}
\arguments{
\item{file_name}{directory name}

\item{ext_name}{template name}
}
\description{
Creates a new GeoX project folder
}
\examples{
# use_quarto_ext('my_new_project')
}
