% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalise.R
\name{regionalise_ga}
\alias{regionalise_ga}
\title{Regionalise a Google Analytics dataset with lat long columns}
\usage{
regionalise_ga(ga, test_region, control_region, micro = FALSE)
}
\arguments{
\item{control_region}{}
}
\description{
Regionalise a Google Analytics dataset with lat long columns
}
