% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalise.R
\name{uplift_ga}
\alias{uplift_ga}
\title{Title}
\usage{
uplift_ga(
  ga = NULL,
  test_regions = NULL,
  exclude_regions = NULL,
  test_start = NULL,
  test_end = NULL,
  benchmark_advertiser = NA,
  benchmark_source = NA,
  benchmark_comment = NA,
  benchmark_schedule = NA,
  auto_regionalise = TRUE
)
}
\arguments{
\item{auto_regionalise}{}
}
\description{
Title
}
