% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_template.R
\name{use_quarto_ext}
\alias{use_quarto_ext}
\title{Creates a new GeoX project folder}
\usage{
use_quarto_ext(file_name = NULL, ext_name = "quartotemplate")
}
\arguments{
\item{file_name}{directory name}

\item{ext_name}{template name}
}
\description{
Creates a new GeoX project folder
}
\examples{
# use_quarto_ext('my_new_project')
}
